refactor: App과 비슷한 역할을 하는 MovieDataLoader을 App과 병합
